deleted old files, replaced new ones, added screenshots, and wrote README
I realized I read the assignment instructions wrong and only needed to put files in that were related to project 1 and 2. So, I deleted the old files and replaced them with the relevant ones. In addition, I added the relevant screenshots to the projects and answered the questions in the README file as asked.